commit 3: added appointment and doctor classes (they and patient was moved to subjects folder)
