+++ Polyclinic/Polyclinic/Enums/DoctorSpecialization.cs
@@ -1,0 +1,7 @@
+namespace Polyclinic.Enums;
+
+public class Specialization
+{
+    public int Id { get; set; }
+    public required string Name { get; set; } // example Therapy, Surgery, Diagnostics, Emergency
+}
