--- Polyclinic/Polyclinic/Program.cs
@@ -1,3 +1,0 @@
-﻿// See https://aka.ms/new-console-template for more information
-
-Console.WriteLine("Hello, World!");
